Unify client and server certificate selection text
I noticed this as I was looking for where we officially wrote down the
implications of the X.509 Key Usage extension. We wrote it down for the
server, but forgot to for the client.

The root issue is that we unified certificate negotiation in TLS 1.3
(ClientHello/Certificate and CertificateRequest/Certificate are now nice
and symmetric), but forgot to correspondingly unify some of this prose.
As a result, we said the same thing in different ways, and forgot
different things in either place.

This change merges the two.